feat: Add table chart type option
- Remove separate files for components used in one place (chart type selector)
- Move label/icon helpers to configurator
