+++ app/charts/index.ts
@@ -130,7 +130,7 @@ export const getPossibleChartType = ({
     (dim) => dim.__typename === "TemporalDimension"
   );
 
-  const catBased: ChartType[] = ["bar", "column", "pie"];
+  const catBased: ChartType[] = ["bar", "column", "pie", "table"];
   const multipleQ: ChartType[] = ["scatterplot"];
   const timeBased: ChartType[] = ["line", "area"];
