+++ app/configurator/components/chart-controls/control-tab.tsx
@@ -2,7 +2,7 @@ import { Trans } from "@lingui/macro";
 import { Button, Flex, Text } from "@theme-ui/components";
 
 import { FieldProps } from "../..";
-import { getFieldLabel, getIconName } from "../../../domain/helpers";
+import { getFieldLabel, getIconName } from "../ui-helpers";
 import { ComponentFieldsFragment } from "../../../graphql/query-hooks";
 import { Icon, IconName } from "../../../icons";
 import { ReactNode } from "react";
