+++ app/configurator/components/chart-type-selector.tsx
@@ -1,13 +1,16 @@
 import { Trans } from "@lingui/macro";
-import { Box, Grid } from "@theme-ui/components";
-import React from "react";
+import { Box, Button, Grid, Text } from "@theme-ui/components";
+import React, { SyntheticEvent } from "react";
 import { getPossibleChartType } from "../../charts";
 import { ChartType, ConfiguratorStateSelectingChartType } from "..";
 import { useDataCubeMetadataWithComponentValuesQuery } from "../../graphql/query-hooks";
 import { useLocale } from "../../locales/use-locale";
 import { SectionTitle } from "./chart-controls/section";
-import { ChartTypeSelectorField } from "./field";
 import { Hint, Loading } from "../../components/hint";
+import { FieldProps, useChartTypeSelectorField } from "../config-form";
+import { DataCubeMetadata } from "../../graphql/types";
+import { Icon } from "../../icons";
+import { getFieldLabel, getIconName } from "./ui-helpers";
 
 const chartTypes: ChartType[] = [
   // "bar",
@@ -16,7 +19,102 @@ const chartTypes: ChartType[] = [
   "area",
   "scatterplot",
   "pie",
+  "table",
 ];
+
+export const ChartTypeSelectionButton = ({
+  label,
+  value,
+  checked,
+  disabled,
+  onClick,
+}: {
+  label: string;
+  disabled?: boolean;
+  onClick: (e: SyntheticEvent<HTMLButtonElement>) => void;
+} & FieldProps) => {
+  return (
+    <Button
+      variant="reset"
+      tabIndex={0}
+      value={value}
+      onClick={onClick}
+      sx={{
+        width: "86px",
+        height: "86px",
+        mt: 4,
+        borderRadius: "default",
+
+        backgroundColor: checked ? "primary" : "monochrome100",
+        color: checked
+          ? "mutedColored"
+          : disabled
+          ? "monochrome500"
+          : "primary",
+
+        display: "flex",
+        flexDirection: "column",
+        justifyContent: "center",
+        alignItems: "center",
+
+        cursor: disabled ? "initial" : "pointer",
+        pointerEvents: disabled ? "none" : "initial",
+
+        transition: "all .2s",
+
+        ":hover": {
+          backgroundColor: disabled
+            ? "mutedColored"
+            : checked
+            ? "primary"
+            : "mutedDarker",
+        },
+      }}
+    >
+      <Icon name={getIconName(label)} />
+      <Text
+        variant="paragraph2"
+        sx={{
+          color: disabled
+            ? "monochrome600"
+            : checked
+            ? "monochrome100"
+            : "monochrome700",
+          fontSize: [2, 2, 2],
+        }}
+      >
+        {getFieldLabel(label)}
+      </Text>
+    </Button>
+  );
+};
+
+const ChartTypeSelectorField = ({
+  label,
+  value,
+  metaData,
+  disabled,
+  ...props
+}: {
+  label: string;
+  value: string;
+  metaData: DataCubeMetadata;
+  disabled?: boolean;
+}) => {
+  const field = useChartTypeSelectorField({
+    value,
+    metaData,
+  });
+
+  return (
+    <ChartTypeSelectionButton
+      disabled={disabled}
+      label={label}
+      {...field}
+    ></ChartTypeSelectionButton>
+  );
+};
+
 export const ChartTypeSelector = ({
   state,
 }: {
